tui: extract external agent migration renderer
